jobdhundo_backend/requirements.txt
Django>=5.0,<6.1
djangorestframework
django-cors-headers
# Production specific
gunicorn
psycopg2-binary
dj-database-url
whitenoise
